feat: wire Tauri commands, tauri-specta bindings, and real app state (Step 5)
- commands/providers.rs: list_providers, test_provider_connection,
  list_models. commands/runs.rs: run_generation (the Playground's Run
  button - calls the provider, estimates cost, persists the Run either
  way, returns it). Shared require_api_key helper in commands/mod.rs.
- lib.rs setup() now resolves the app data dir, opens the real SQLite
  Database, and manages Database/ProviderRegistry/PricingTable - these
  three modules are no longer inert, only-tested library code.
- Wired specta + tauri-specta, pinned to =2.0.0-rc.25 (the only
  Tauri v2-compatible release; still RC after 25 candidates - user
  explicitly chose this over hand-written TS types). Generates
  src/lib/bindings.ts from the same command list used at runtime.
- Refactored AppError to a plain #[serde(tag/content)] derive (dropped
  the hand-rolled Serialize impl) so specta::Type derives too; call
  sites now pre-format full messages instead of relying on a generic
  Display prefix (audited - every call site already did this).
- Fixed a real panic caught by the tauri dev smoke test: specta
  forbids exporting i64/u64 to TypeScript. RunId/ExperimentId now
  cross IPC as strings (still i64 internally); RunResult.duration_ms/
  ttft_ms switched from u64 to u32.
- Database::with_connection now uses tauri::async_runtime::spawn_blocking
  (per the original architecture decision, applied now that real async
  commands call it). runs_repo::insert_run/get_run are async;
  insert_run takes NewRun by value.

Verified: tauri dev launches cleanly (confirmed by user), bindings.ts
generated correctly, cargo check/clippy/fmt/test (23 passed, 1 ignored)
and npm run build all clean.

+++ src-tauri/src/providers/openai.rs
@@ -265,7 +265,7 @@ impl LlmProvider for OpenAiProvider {
             .json()
             .await
             .map_err(|e| AppError::Provider(format!("unexpected OpenAI response: {e}")))?;
-        let duration_ms = started.elapsed().as_millis() as u64;
+        let duration_ms = started.elapsed().as_millis() as u32;
 
         let text = body
             .choices
